AI-server/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langgraph
langchain-chroma
psycopg2-binary
redis
python-dotenv
pandas
sqlalchemy
pydantic-settings
